#9 fix refresh problem on interval item changes
when adding or removing items, it can happen that the placeholder icon
for the value labels is shown, although there is enough space and vice
versa. To overcome this situation, angular events are used that are
fired every time a interval item is deleted or added. Emitting and
broadcasting those events guarantee that the visibilty of the label is
always correct.