#9 fix wrong positioning of interval values popup
when the width of an indicator interval is to small to display all value
labels, an icon with an tooltip is displayed instead. When trying to
edit the values of such an interval opens the popup at a very wrong
place. The reason is that popover is bound to the invisble div that
contains the value labels. To overcome this the popup is now bound to a
seperate div that wraps the value label div and icon div.