#45 corrected tests for table directives
the tests suites for the table directive contains tests that checks if
the warn badges are correctly shown. Those tests check if the html
representation of the compiled directive contains the respective html of
the warn badge. since phantomjs escapes signs such as '>' and '>' those
tests fail. furhtermore the those tests are uneffective since ng-show
and ng-hide are used instead of ng-if which removes the parts
effectively from the dom.